program added

+++ Practice_point/c_notes/float_double.c
@@ -2,11 +2,12 @@
 #include<math.h>
 int main()
 {
-	float x=4.0,y;
-//	double x=4,y;
-	double y1;
-	y1=sqrt(x);
-	if(y1==2)
+//	float x = 4.0, y;
+	double x = 4, y;
+//	double y1;
+	float y1;
+	y1 = sqrt(x);
+	if (y1 == 2)
 		printf("Equal\n");
 	else
 		printf("Not equal\n");
